agent/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-groq>=0.2.0
yfinance>=0.2.40
tavily-python>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
grandalf>=0.4
reportlab>=4.0